Class "Line.Transect.Design" extends Class "Survey.Design"

Class "Line.Transect.Design" is an S4 class detailing the type of line transect design.

Slots

line.length

Numeric value defining the total line length to be generated (may be multiple values relating to each stratum).

bounding.shape

relevant for zigzag designs, either a minimum bounding "rectangle" or a "convex.hull".

Methods

generate.transects

signature=(object = "Line.Transect.Design", quiet = FALSE, ...): generates a set of transects from the design.

See also