Class "Segment.Transect.Design" extends Class "Survey.Design"

Class "Segment.Transect.Design" is an S4 class detailing the a segmented line transect design.

Slots

seg.length

length of the transect segment.

seg.threshold

this is a percentage threshold value applicable to segmented grid designs controlling which partial segments are discarded around the survey region boundary. By default, the value of 50, means that only segments that are more than half inside the survey region will be retained. To retain all segments, no matter how small they are when clipped to the survey region boundary set this value to 0.

offset

a value to offset a return transect by so segments become pairs of segments (not yet implemented).

Methods

generate.transects

signature=(object = "Line.Transect.Design", quiet = FALSE, ...): generates a set of transects from the design.

See also