Virtual Class "Survey.Design"

Virtual Class "Survey.Design" is an S4 class detailing the survey design.

Slots

region

An object of class 'Region' defining the study area.

design

Character value describing the name of the design.

samplers

Numeric values defining the number of samplers in each stratum.

effort.allocation

numeric values used to indicate the proportion of effort to be allocated to each strata from number of samplers or line length. If length 0, effort allocated based on stratum area.

spacing

used by systematic designs, numeric value to define spacing between transects.

design.angle

numeric value detailing the angle of the design. Can provide multiple values relating to strata. The use of the angle varies with design, it can be either the angle of the grid of points, the angle of lines or the design axis for the zigzag design.

edge.protocol

Character value defining whether a "minus" or "plus" sampling strategy should be used.

truncation

Object of class "numeric"; The maximum distance at which observations can be made. This is used to determine the covered area during the coverage calculations.

coverage.grid

The coverage grid used to assess the uniformity of coverage during simulations.

coverage.scores

The average number of times each point in the coverage grid is included in a survey.

coverage.reps

The number of times the coverage simulation was repeated.

design.statistics

A list of values obtained when investigating coverage. This includes the minimum, maximum, mean and median

Methods

generate.transects

signature 'Survey.Design': Generates a set of transects from the design.

plot

signature 'Survey.Design,ANY': Plots the coverage scores contained within an object of class 'Survey.Design' and provides a colour key relating to the coverage scores. This allows the user to assess how even the coverage is across the survey region.

show

signature 'Survey.Design': Gives a summary of the design description, stratum areas and coverage scores if the coverage simulation has been run on the design. The coverage score summary details the minimum, maximum, mean and medium coverage scores across the study region. Also gives summaries of other design measures such as the number of samplers, line length, trackline length, cyclic trackline length, covered area and percentage of region covered.

See also