Distance Sampling Survey Design (dssd)

Creates survey designs for distance sampling surveys. These designs can be assessed for various effort and coverage statistics. Once the user is satisfied with the design characteristics they can generate a set of transects to use in their distance sampling survey. Many of the designs implemented in this R package were first made available in our 'Distance' for Windows software and are detailed in Chapter 7 of Advanced Distance Sampling, Buckland et. al. (2008, ISBN-13: 978-0199225873).

Details

The main functions in this package are: make.region, make.design, generate.transects and run.coverage. See also write.transects for examples of how to export surveys generated by dssd.

Further information on distance sampling methods and example code is available at https://distancesampling.org/R/.

Also see our website for vignettes / example code at https://distancesampling.org/resources/vignettes.html.

For help with distance sampling and this package, there is a Google Group https://groups.google.com/forum/#!forum/distance-sampling.

Author

Laura Marshall <lhm@st-and.ac.uk>