Plot

Plot's an object of class Region with optionally a set of transects or the points of a coverage grid if supplied as the y argument.

Usage

# S4 method for class 'Region,ANY'
plot(
  x,
  y,
  main = "",
  region.col = "default",
  strata = "all",
  line.col = gray(0.2),
  legend.params = list()
)

# S4 method for class 'Region,Transect'
plot(
  x,
  y,
  main = "",
  region.col = "default",
  strata = "all",
  line.col = gray(0.2),
  col = "blue",
  lwd = 1,
  covered.area = FALSE,
  legend.params = list()
)

# S4 method for class 'Region,Coverage.Grid'
plot(
  x,
  y,
  main = "",
  region.col = "default",
  strata = "all",
  line.col = gray(0.2),
  col = "black",
  cex = 1
)

Arguments

x

object of class Region or inheriting from Survey

y

an object inheriting from class Transect or an object of class Coverage.Grid

main

the main title for the plot

region.col

colours for the strata

strata

the strata name or number to be plotted. By default all strata will be plotted.

line.col

sets the line colour for the lines around the survey region.

legend.params

deprecated since implementation of ggplot2

col

sets the colour of the transects / coverage grid points

lwd

sets the line width of the transects

covered.area

boolean value saying whether the covered area should be plotted.

cex

affects the size of the point on the coverage grid

Value

returns a ggplot object